+++ CHANGELOG.md
@@ -24,6 +24,9 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Record cluster and worker thread scrape failures in internal histograms,
+  including timeouts, worker-reported errors, and failures with no known worker errors.
+
 ## [0.16.0] - 2026-08-24
 
 This release marks our first release as a Prometheus subproject.